WORKS! but
after entering the Game Scene two errors (or more, but 2 per player joining) will apper:
MissingReferenceException: The object of type 'UnityEngine.UI.Button' has been destroyed but you are still trying to access it.
Your script should either check if it is null or you should not destroy the object.

MissingReferenceException: The object of type 'UnityEngine.RectTransform' has been destroyed but you are still trying to access it.
Your script should either check if it is null or you should not destroy the object.

those are fine, they happen cause we change scenes too fast, and it still tries to update the lobby info for the joining player.
I will try to fix it, but it's not as importnat right now